backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
httpx==0.28.1
supabase==2.15.2
python-dotenv==1.1.0
pydantic==2.11.3
slowapi==0.1.9
python-multipart==0.0.20
scikit-learn==1.6.1
joblib==1.5.0
numpy==2.2.5
